deleted unused wildfire code
