+++ crawlers/earthquake/earthquake.log
@@ -85,3 +85,7 @@
 2024-04-28 01:55:54,892 - INFO - New earthquake data appended to earthquake.json
 2024-04-28 03:56:51,192 - INFO - New earthquake data appended to earthquake.json
 2024-04-29 01:36:18,368 - INFO - New earthquake data appended to earthquake.json
+2024-04-30 00:09:23,733 - INFO - New earthquake data appended to earthquake.json
+2024-04-30 00:42:58,546 - INFO - New earthquake data appended to earthquake.json
+2024-04-30 00:44:12,875 - INFO - New earthquake data appended to earthquake.json
+2024-04-30 00:44:55,166 - INFO - New earthquake data appended to earthquake.json
