+++ crawlers/weather_gov/weather_gov.log
@@ -5,3 +5,4 @@
 2024-04-28 20:23:24,358 - INFO - New weather alert data appended to new_weather_gov_data.json
 2024-04-28 20:36:54,850 - INFO - New weather alert data appended to new_weather_gov_data.json
 2024-04-29 01:35:48,241 - INFO - New weather alert data appended to new_weather_gov_data.json
+2024-04-30 00:44:58,415 - INFO - New weather alert data appended to new_weather_gov_data.json
