+++ crawlers/wildfire/fires_ca.log
@@ -1,0 +1,7 @@
+2023-11-30 22:41:58,898 - ERROR - An error occurred: 'metadata'
+2023-11-30 22:45:16,509 - INFO - New fire data appended to fires_ca.json
+2023-12-15 21:30:42,044 - INFO - New fire data appended to fires_ca.json
+2024-04-29 23:55:06,034 - INFO - New fire data appended to fires_ca.json
+2024-04-29 23:57:58,379 - INFO - New fire data appended to fires_ca.json
+2024-04-30 00:09:02,708 - INFO - New fire data appended to fires_ca.json
+2024-04-30 00:45:02,282 - INFO - New fire data appended to fires_ca.json
