--- crawlers/wildfire/fires_ca/fires_ca.log
@@ -1,3 +1,0 @@
-2023-11-30 22:41:58,898 - ERROR - An error occurred: 'metadata'
-2023-11-30 22:45:16,509 - INFO - New fire data appended to fires_ca.json
-2023-12-15 21:30:42,044 - INFO - New fire data appended to fires_ca.json
