--- crawlers/wildfire/inciweb/news.csv
@@ -1,4 +1,0 @@
-,URL,Current as of,Incident Type,Cause,Date of Origin,Location,Coordinates,Size,Percent of Perimeter Contained
-0,https://inciweb.nwcg.gov//incident-information/caknp-redwood-fire,"Sat, 11/25/2023 - 11:48",Wildfire,Lightning,"Tue, 08/15/2023 - 00:00",Sequoia National Park,"36°                                  32'                                  Latitude
-                                  
-                                 -118°                                  38'                                  Longitude","2,248 Acres",70%
