--- crawlers/wildfire/inciweb/requirements.txt
@@ -1,99 +1,0 @@
-anyio==4.0.0
-appnope==0.1.3
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
-asttokens==2.4.0
-async-lru==2.0.4
-attrs==23.1.0
-Babel==2.13.0
-backcall==0.2.0
-beautifulsoup4==4.12.2
-bleach==6.1.0
-certifi==2023.7.22
-cffi==1.16.0
-charset-normalizer==3.3.1
-comm==0.1.4
-debugpy==1.8.0
-decorator==5.1.1
-defusedxml==0.7.1
-exceptiongroup==1.1.3
-executing==2.0.0
-fastjsonschema==2.18.1
-fqdn==1.5.1
-idna==3.4
-importlib-metadata==6.8.0
-ipykernel==6.25.2
-ipython==8.16.1
-isoduration==20.11.0
-jedi==0.19.1
-Jinja2==3.1.2
-json5==0.9.14
-jsonpointer==2.4
-jsonschema==4.19.1
-jsonschema-specifications==2023.7.1
-jupyter-events==0.8.0
-jupyter-lsp==2.2.0
-jupyter_client==8.4.0
-jupyter_core==5.4.0
-jupyter_server==2.8.0
-jupyter_server_terminals==0.4.4
-jupyterlab==4.0.7
-jupyterlab-pygments==0.2.2
-jupyterlab_server==2.25.0
-MarkupSafe==2.1.3
-matplotlib-inline==0.1.6
-mistune==3.0.2
-nbclient==0.8.0
-nbconvert==7.9.2
-nbformat==5.9.2
-nest-asyncio==1.5.8
-notebook==7.0.6
-notebook_shim==0.2.3
-numpy==1.26.1
-overrides==7.4.0
-packaging==23.2
-pandas==2.1.1
-pandocfilters==1.5.0
-parso==0.8.3
-pexpect==4.8.0
-pickleshare==0.7.5
-platformdirs==3.11.0
-prometheus-client==0.17.1
-prompt-toolkit==3.0.39
-psutil==5.9.6
-ptyprocess==0.7.0
-pure-eval==0.2.2
-pycparser==2.21
-Pygments==2.16.1
-python-dateutil==2.8.2
-python-json-logger==2.0.7
-pytz==2023.3.post1
-PyYAML==6.0.1
-pyzmq==25.1.1
-referencing==0.30.2
-requests==2.31.0
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-rpds-py==0.10.6
-Send2Trash==1.8.2
-six==1.16.0
-sniffio==1.3.0
-soupsieve==2.5
-stack-data==0.6.3
-tabulate==0.9.0
-terminado==0.17.1
-tinycss2==1.2.1
-tomli==2.0.1
-tornado==6.3.3
-traitlets==5.11.2
-types-python-dateutil==2.8.19.14
-typing_extensions==4.8.0
-tzdata==2023.3
-uri-template==1.3.0
-urllib3==2.0.7
-wcwidth==0.2.8
-webcolors==1.13
-webencodings==0.5.1
-websocket-client==1.6.4
-zipp==3.17.0
