--- crawlers/wildfire/inciweb/wildfire.json
@@ -1,6 +1,0 @@
-[
-    {
-        "URL": "https://example.com",
-        "Timestamp": "2023-11-18 17:18:28"
-    }
-]
