+++ embeddings.py
@@ -21,7 +21,7 @@ def create_store_eq1_embeddings():
         embedding=OpenAIEmbeddings(),
         ids=ids,
     )
-    print("data inserted")
+    print("earthquake data inserted", len(documents))
     return vectordb
 
 
@@ -40,7 +40,26 @@ def create_store_we_embeddings():
         embedding=OpenAIEmbeddings(),
         ids=ids,
     )
-    print("data inserted")
+    print("weather data inserted", len(documents))
+    return vectordb
+
+
+def create_store_wf_embeddings():
+    loader = JSONLoader(
+        "./crawlers/wildfire/fires_ca.json",
+        jq_schema=".features[]",
+        text_content=False,
+    )
+    documents = loader.load()
+    ids = [get_id_from_document("wildfire", doc) for doc in documents]
+
+    vectordb = Chroma.from_documents(
+        client=chroma_client,
+        documents=documents,
+        embedding=OpenAIEmbeddings(),
+        ids=ids,
+    )
+    print("wildfire data inserted", len(documents))
     return vectordb
 
 
@@ -50,3 +69,5 @@ if __name__ == "__main__":
     print("created earthquake embeddings")
     create_store_we_embeddings()
     print("created weather embeddings")
+    create_store_wf_embeddings()
+    print("created wildfire embeddings")
